Develop

+++ Convertidor/Services/Rh/RhDocumentosService.cs
@@ -10,17 +10,20 @@ namespace Convertidor.Data.Repository.Rh
         private readonly IRhDescriptivasService _descriptivaService;
         private readonly IRhPersonasRepository _personasRepository;
         private readonly ISisUsuarioRepository _sisUsuarioRepository;
-   
+        private readonly IRhDocumentosDetallesService _documentosDetallesService;
+
 
         public RhDocumentosService(IRhDocumentosRepository repository, 
                                         IRhDescriptivasService descriptivaService,
                                         IRhPersonasRepository personasRepository,
-                                        ISisUsuarioRepository sisUsuarioRepository)
+                                        ISisUsuarioRepository sisUsuarioRepository,
+                                        IRhDocumentosDetallesService documentosDetallesService)
         {
             _repository = repository;
             _descriptivaService = descriptivaService;
             _personasRepository = personasRepository;
             _sisUsuarioRepository = sisUsuarioRepository;
+            _documentosDetallesService = documentosDetallesService;
         }
        
         public async Task<List<RhDocumentosResponseDto>> GetByCodigoPersona(int codigoPersona)
@@ -53,16 +56,43 @@ namespace Convertidor.Data.Repository.Rh
                 itemResult.CodigoPersona = dtos.CODIGO_PERSONA;
                 itemResult.CodigoDocumento = dtos.CODIGO_DOCUMENTO;
                 itemResult.TipodocumentoId = dtos.TIPO_DOCUMENTO_ID;
+                itemResult.DescripcionDocumento = "";
+                if (dtos.TIPO_DOCUMENTO_ID != null)
+                {
+                    var descriptivaDocumento=await _descriptivaService.GetByCodigoDescriptiva(itemResult.TipodocumentoId);
+                    if (descriptivaDocumento != null)
+                    {
+                        itemResult.DescripcionDocumento = descriptivaDocumento.DESCRIPCION;
+                    }
+                }
+                
+              
                 itemResult.NumeroDocumento = dtos.NUMERO_DOCUMENTO;
                 itemResult.FechaVencimiento = dtos.FECHA_VENCIMIENTO;
                 itemResult.FechaVencimientoString =Fecha.GetFechaString(dtos.FECHA_VENCIMIENTO);
                 FechaDto fechaVencimientoObj = Fecha.GetFechaDto(dtos.FECHA_VENCIMIENTO);
                 itemResult.FechaVencimientoObj = (FechaDto)fechaVencimientoObj;
                 itemResult.TipoGradoid = dtos.TIPO_GRADO_ID;
+                itemResult.DescripcionTipoGrado = "";
+                if (dtos.TIPO_GRADO_ID != null)
+                {
+                    var descriptivaTipoGrado=await _descriptivaService.GetByCodigoDescriptiva(itemResult.TipoGradoid);
+                    if (descriptivaTipoGrado != null)
+                    {
+                        itemResult.DescripcionTipoGrado = descriptivaTipoGrado.DESCRIPCION;
+                    }
+                }
                 itemResult.GradoId = dtos.GRADO_ID;
-                itemResult.Extra1 = dtos.EXTRA1;
-                itemResult.Extra2 = dtos.EXTRA2;
-                itemResult.Extra3 = dtos.EXTRA3; 
+                itemResult.DescripcionTipoGrado= "";
+                if (dtos.TIPO_GRADO_ID != null)
+                {
+                    var descriptivaGrado=await _descriptivaService.GetByCodigoDescriptiva(itemResult.GradoId);
+                    if (descriptivaGrado != null)
+                    {
+                        itemResult.DescripcionTipoGrado = descriptivaGrado.DESCRIPCION;
+                    }
+                }
+               
                 
              
           
@@ -115,7 +145,7 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-                var tipoDocumento = await _descriptivaService.GetByTitulo(26);
+                var tipoDocumento = await _descriptivaService.GetByCodigoDescriptiva(dto.TipoDocumentoId);
                 if (tipoDocumento==null)
                 {
                     result.Data = null;
@@ -124,7 +154,7 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-                if(dto.NumeroDocumento == null) 
+                if(string.IsNullOrEmpty(dto.NumeroDocumento)) 
                 {
                     result.Data = null;
                     result.IsValid = false;
@@ -132,54 +162,40 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-                var tipoGradoId = await _descriptivaService.GetByTitulo(30);
-                if (tipoGradoId == null)
+                if (dto.TipoGradoId != null && dto.TipoGradoId > 0)
                 {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Tipo Grado Id Documento Invalido";
-                    return result;
+                    var tipoGrado = await _descriptivaService.GetByCodigoDescriptiva(dto.TipoGradoId);
+                    if (tipoGrado == null)
+                    {
+                        result.Data = null;
+                        result.IsValid = false;
+                        result.Message = "Tipo Grado Id Documento Invalido";
+                        return result;
+                    }
+
                 }
 
-                var GradoId = await _descriptivaService.GetByTitulo(31);
-                if (GradoId == null)
+                if (dto.GradoId==null && dto.GradoId > 0)
                 {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Grado Id Invalido";
-                    return result;
+                    var grado = await _descriptivaService.GetByCodigoDescriptiva(dto.GradoId);
+                    if (grado == null)
+                    {
+                        result.Data = null;
+                        result.IsValid = false;
+                        result.Message = "Grado Id Invalido";
+                        return result;
+                    }
                 }
-                if (dto.Extra1 is not null && dto.Extra1.Length > 100)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra1 Invalido";
-                    return result;
-                }
-                if (dto.Extra2 is not null && dto.Extra2.Length > 100)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra2 Invalido";
-                    return result;
-                }
-
-                if (dto.Extra3 is not null && dto.Extra3.Length > 100)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra3 Invalido";
-                    return result;
-                }
-
-
-
-
+                
                 documento.CODIGO_PERSONA = dto.CodigoPersona;
                 documento.CODIGO_DOCUMENTO = dto.CodigoDocumento;
                 documento.TIPO_DOCUMENTO_ID = dto.TipoDocumentoId;
                 documento.NUMERO_DOCUMENTO = dto.NumeroDocumento;
-                documento.FECHA_VENCIMIENTO = dto.FechaVencimiento;
+                if (dto.FechaVencimiento != null)
+                {
+                    documento.FECHA_VENCIMIENTO = (DateTime)dto.FechaVencimiento;
+                }
+           
                 documento.TIPO_GRADO_ID = dto.TipoGradoId;
                 documento.GRADO_ID = dto.GradoId;
               
@@ -228,16 +244,8 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-                var tipoDocumento = await _descriptivaService.GetByTitulo(26);
-                if (tipoDocumento == null)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Tipo Documento  Invalido";
-                    return result;
-                }
-
-                if (dto.NumeroDocumento == null)
+             
+                if(string.IsNullOrEmpty(dto.NumeroDocumento)) 
                 {
                     result.Data = null;
                     result.IsValid = false;
@@ -245,52 +253,54 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-                var tipoGradoId = await _descriptivaService.GetByTitulo(30);
-                if (tipoGradoId == null)
+                var tipoDocumento = await _descriptivaService.GetByCodigoDescriptiva(dto.TipoDocumentoId);
+                if (tipoDocumento==null)
                 {
                     result.Data = null;
                     result.IsValid = false;
-                    result.Message = "Tipo Grado Id Documento Invalido";
+                    result.Message = "Tipo Documento  Invalido";
                     return result;
                 }
 
-                var Grado = await _descriptivaService.GetByTitulo(31);
-                if (Grado == null)
+           
+
+                if (dto.TipoGradoId != null && dto.TipoGradoId > 0)
                 {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Grado Id Documento Invalido";
-                    return result;
-                }
-                if (dto.Extra1 is not null && dto.Extra1.Length > 100)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra1 Invalido";
-                    return result;
-                }
-                if (dto.Extra2 is not null && dto.Extra2.Length > 100)
-                {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra2 Invalido";
-                    return result;
+                    var tipoGrado = await _descriptivaService.GetByCodigoDescriptiva(dto.TipoGradoId);
+                    if (tipoGrado == null)
+                    {
+                        result.Data = null;
+                        result.IsValid = false;
+                        result.Message = "Tipo Grado Id Documento Invalido";
+                        return result;
+                    }
+
                 }
 
-                if (dto.Extra3 is not null && dto.Extra3.Length > 100)
+                if (dto.GradoId==null && dto.GradoId > 0)
                 {
-                    result.Data = null;
-                    result.IsValid = false;
-                    result.Message = "Extra3 Invalido";
-                    return result;
+                    var grado = await _descriptivaService.GetByCodigoDescriptiva(dto.GradoId);
+                    if (grado == null)
+                    {
+                        result.Data = null;
+                        result.IsValid = false;
+                        result.Message = "Grado Id Invalido";
+                        return result;
+                    }
                 }
+
+              
 
                 RH_DOCUMENTOS entity = new RH_DOCUMENTOS();
                 entity.CODIGO_DOCUMENTO = await _repository.GetNextKey();
                 entity.CODIGO_PERSONA = dto.CodigoPersona;
                 entity.TIPO_DOCUMENTO_ID = dto.TipoDocumentoId;
                 entity.NUMERO_DOCUMENTO = dto.NumeroDocumento;
-                entity.FECHA_VENCIMIENTO = dto.FechaVencimiento;
+                if (dto.FechaVencimiento != null)
+                {
+                    entity.FECHA_VENCIMIENTO = (DateTime)dto.FechaVencimiento;
+                }
+             
                 entity.TIPO_GRADO_ID = dto.TipoGradoId;
                 entity.GRADO_ID = dto.GradoId;
                 
@@ -357,7 +367,15 @@ namespace Convertidor.Data.Repository.Rh
                     return result;
                 }
 
-
+                var detalleDocumento = await _documentosDetallesService.GetByDocumento(dto.CodigoDocumento);
+                if (detalleDocumento != null && detalleDocumento.Count>0)
+                {
+                    result.Data = null;
+                    result.IsValid = false;
+                    result.Message = "No puede eliminar un Documento que cotiene detalle";
+                    return result;
+                }
+                
                 var deleted = await _repository.Delete(dto.CodigoDocumento);
 
                 if (deleted.Length > 0)
